requirements.txt
flask==2.3.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
pyjwt==2.8.0
bcrypt==4.1.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pydantic==2.5.0
boto3==1.34.0
